monitoring/telegram-bot/requirements.txt
python-telegram-bot==13.15
fastapi==0.109.2
uvicorn==0.27.1 